Changed UI and created theme. Also updated the recipe
Created a UI on figma which I will refer back to later. Theme is created, but needs to be updated as I didn't understand how it worked in the beginning. Recipe now updates and has a black color if the inventory has the items needed to make that recipe. Will update all the ui later to make it more easily readable